refactor: reorganise files in runtime pkg, add shared utils in kernel
