+++ packages/kernel/src/utilities.ts
@@ -7,14 +7,71 @@ import { Metadata } from '@aurelia/metadata';
 /** @internal */ export const hasMetadata = Metadata.has;
 /** @internal */ export const defineMetadata = Metadata.define;
 
-/** @internal */ export const isPromise = <T>(v: unknown): v is Promise<T> => v instanceof Promise;
+/**
+ * Returns true if the value is a Promise via checking if it's an instance of Promise.
+ * This does not work for objects across different realms (e.g., iframes).
+ * An utility to be shared among core packages for better size optimization
+ */
+export const isPromise = <T>(v: unknown): v is Promise<T> => v instanceof Promise;
 
+/**
+ * Returns true if the value is an Array via checking if it's an instance of Array.
+ * This does not work for objects across different realms (e.g., iframes).
+ * An utility to be shared among core packages for better size optimization
+ */
+export const isArray = <T>(v: unknown): v is T[] => v instanceof Array;
+
+/**
+ * Returns true if the value is a Set via checking if it's an instance of Set.
+ * This does not work for objects across different realms (e.g., iframes).
+ * An utility to be shared among core packages for better size optimization
+ */
+export const isSet = <T>(v: unknown): v is Set<T> => v instanceof Set;
+
+/**
+ * Returns true if the value is a Map via checking if it's an instance of Map.
+ * This does not work for objects across different realms (e.g., iframes).
+ * An utility to be shared among core packages for better size optimization
+ */
+export const isMap = <T, K>(v: unknown): v is Map<T, K> => v instanceof Map;
+
+/**
+ * Returns true if the value is an object via checking if it's an instance of Object.
+ * This does not work for objects across different realms (e.g., iframes).
+ * An utility to be shared among core packages for better size optimization
+ */
+export const isObject = (v: unknown): v is object => v instanceof Object;
+
+/**
+ * Returns true if the value is a function
+ * An utility to be shared among core packages for better size optimization
+ */
 // eslint-disable-next-line @typescript-eslint/ban-types
-/** @internal */ export const isFunction = <T extends Function>(v: unknown): v is T => typeof v === 'function';
+export const isFunction = <T extends Function>(v: unknown): v is T => typeof v === 'function';
 
-/** @internal */ export const isString = (v: unknown): v is string => typeof v === 'string';
-/** @internal */ export const createObject = <T extends object>() => Object.create(null) as T;
-/** @internal */ export const createError = (message: string) => new Error(message);
+/**
+ * Returns true if the value is a string
+ * An utility to be shared among core packages for better size optimization
+ */
+export const isString = (v: unknown): v is string => typeof v === 'string';
+
+/**
+ * Returns true if the value is a symbol
+ * An utility to be shared among core packages for better size optimization
+ */
+export const isSymbol = (v: unknown): v is string => typeof v === 'symbol';
+
+/**
+ * Returns true if the value is a number
+ * An utility to be shared among core packages for better size optimization
+ */
+export const isNumber = (v: unknown): v is number => typeof v === 'number';
+
+/**
+ * Create an object with no prototype to be used as a record
+ * An utility to be shared among core packages for better size optimization
+ */
+export const createRecord = <T>() => Object.create(null) as Record<string, T>;
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export type AnyFunction = (...args: any) => any;
