+++ packages/runtime/src/proxy-decorators.ts
@@ -1,5 +1,5 @@
 import { Class, Constructable } from '@aurelia/kernel';
-import { defineHiddenProp, safeString } from '../utilities';
+import { defineHiddenProp, safeString } from './utilities';
 import { nowrapClassKey, nowrapPropKey } from './proxy-observation';
 
 export function nowrap(): (target: unknown, context: ClassDecoratorContext | ClassFieldDecoratorContext) => void;
